osd/OSDMap: do not encode crc bits if crc_full is not valid anymore
we introduced the crc fields for full map and inc map in OSDMap encoding
v8. but in a previous fix, we re-encode the maps if a features bit is
missing when sending MOSDMap message, so on the recipient side, the
re-encoded v8 inc map still carries the invalid full_crc, and the
decoder of inc map assumes that all v8 inc maps have_crc. so it asks for
full map nevertheless. so we need to re-encode the inc map with a lower
encoder version than v8 if we found that the encode_feature does not
match with the one passed in: we are re-encoding the inc map with
different feature bits.

Fixes: http://tracker.ceph.com/issues/17534
Signed-off-by: Kefu Chai <kchai@redhat.com>